renamed views_2 back to views
