+++ api/data_refinery_api/views/__init__.py
@@ -1,0 +1,38 @@
+from data_refinery_api.views.api_token import APITokenView, CreateAPITokenView
+from data_refinery_api.views.compendium_result import (
+    CompendiumResultDetailView,
+    CompendiumResultListView,
+)
+from data_refinery_api.views.computational_result import (
+    ComputationalResultDetailView,
+    ComputationalResultListView,
+)
+from data_refinery_api.views.computed_file import ComputedFileDetailView, ComputedFileListView
+from data_refinery_api.views.dataset import CreateDatasetView, DatasetView
+from data_refinery_api.views.experiment import ExperimentDetailView, ExperimentListView
+from data_refinery_api.views.experiment_document import ExperimentDocumentView
+from data_refinery_api.views.institution import InstitutionListView
+from data_refinery_api.views.jobs import (
+    DownloaderJobDetailView,
+    DownloaderJobListView,
+    ProcessorJobDetailView,
+    ProcessorJobListView,
+    SurveyJobDetailView,
+    SurveyJobListView,
+)
+from data_refinery_api.views.organism import OrganismDetailView, OrganismListView
+from data_refinery_api.views.original_file import OriginalFileListView, OriginalFileDetailView
+from data_refinery_api.views.platform import PlatformListView
+from data_refinery_api.views.processor import ProcessorDetailView, ProcessorListView
+from data_refinery_api.views.sample import SampleListView, SampleDetailView
+from data_refinery_api.views.stats import (
+    AboutStats,
+    FailedDownloaderJobStats,
+    FailedProcessorJobStats,
+    Stats,
+)
+from data_refinery_api.views.transcriptome_index import (
+    TranscriptomeIndexListView,
+    TranscriptomeIndexDetailView,
+)
+from data_refinery_api.views.qn_targets import QNTargetsAvailable, QNTargetsDetailView
