+++ api/data_refinery_api/views/compendium_result.py
@@ -11,14 +11,14 @@ from django_filters.rest_framework import DjangoFilterBackend
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 
-from data_refinery_api.views_2.organism import OrganismSerializer
-from data_refinery_api.views_2.relation_serializers import (
+from data_refinery_api.views.organism import OrganismSerializer
+from data_refinery_api.views.relation_serializers import (
     ComputationalResultNoFilesRelationSerializer,
 )
 
 from data_refinery_common.models import APIToken, CompendiumResult, ComputedFile
 
-from data_refinery_api.views_2.relation_serializers import (
+from data_refinery_api.views.relation_serializers import (
     ComputedFileRelationSerializer,
     ComputedFileWithUrlRelationSerializer,
 )
