+++ api/data_refinery_api/views/computational_result.py
@@ -8,7 +8,7 @@ from django_filters.rest_framework import DjangoFilterBackend
 
 from data_refinery_common.models import APIToken, ComputationalResult
 
-from data_refinery_api.views_2.relation_serializers import (
+from data_refinery_api.views.relation_serializers import (
     ProcessorRelationSerializer,
     OrganismIndexRelationSerializer,
     ComputedFileRelationSerializer,
