+++ api/data_refinery_api/views/experiment.py
@@ -9,7 +9,7 @@ from django_filters.rest_framework import DjangoFilterBackend
 
 from data_refinery_common.models import Experiment, ExperimentAnnotation
 
-from data_refinery_api.views_2.relation_serializers import DetailedExperimentSampleSerializer
+from data_refinery_api.views.relation_serializers import DetailedExperimentSampleSerializer
 
 
 class ExperimentAnnotationSerializer(serializers.ModelSerializer):
