+++ api/data_refinery_api/views/jobs/__init__.py
@@ -1,0 +1,9 @@
+from data_refinery_api.views.jobs.downloader_job import (
+    DownloaderJobDetailView,
+    DownloaderJobListView,
+)
+from data_refinery_api.views.jobs.processor_job import (
+    ProcessorJobDetailView,
+    ProcessorJobListView,
+)
+from data_refinery_api.views.jobs.survey_job import SurveyJobDetailView, SurveyJobListView
