--- api/data_refinery_api/views_2/__init__.py
@@ -1,38 +1,0 @@
-from data_refinery_api.views_2.api_token import APITokenView, CreateAPITokenView
-from data_refinery_api.views_2.compendium_result import (
-    CompendiumResultDetailView,
-    CompendiumResultListView,
-)
-from data_refinery_api.views_2.computational_result import (
-    ComputationalResultDetailView,
-    ComputationalResultListView,
-)
-from data_refinery_api.views_2.computed_file import ComputedFileDetailView, ComputedFileListView
-from data_refinery_api.views_2.dataset import CreateDatasetView, DatasetView
-from data_refinery_api.views_2.experiment import ExperimentDetailView, ExperimentListView
-from data_refinery_api.views_2.experiment_document import ExperimentDocumentView
-from data_refinery_api.views_2.institution import InstitutionListView
-from data_refinery_api.views_2.jobs import (
-    DownloaderJobDetailView,
-    DownloaderJobListView,
-    ProcessorJobDetailView,
-    ProcessorJobListView,
-    SurveyJobDetailView,
-    SurveyJobListView,
-)
-from data_refinery_api.views_2.organism import OrganismDetailView, OrganismListView
-from data_refinery_api.views_2.original_file import OriginalFileListView, OriginalFileDetailView
-from data_refinery_api.views_2.platform import PlatformListView
-from data_refinery_api.views_2.processor import ProcessorDetailView, ProcessorListView
-from data_refinery_api.views_2.sample import SampleListView, SampleDetailView
-from data_refinery_api.views_2.stats import (
-    AboutStats,
-    FailedDownloaderJobStats,
-    FailedProcessorJobStats,
-    Stats,
-)
-from data_refinery_api.views_2.transcriptome_index import (
-    TranscriptomeIndexListView,
-    TranscriptomeIndexDetailView,
-)
-from data_refinery_api.views_2.qn_targets import QNTargetsAvailable, QNTargetsDetailView
