--- api/data_refinery_api/views_2/jobs/__init__.py
@@ -1,9 +1,0 @@
-from data_refinery_api.views_2.jobs.downloader_job import (
-    DownloaderJobDetailView,
-    DownloaderJobListView,
-)
-from data_refinery_api.views_2.jobs.processor_job import (
-    ProcessorJobDetailView,
-    ProcessorJobListView,
-)
-from data_refinery_api.views_2.jobs.survey_job import SurveyJobDetailView, SurveyJobListView
